services/orchestrator/requirements.txt
Flask==2.2.2
google-cloud-pubsub==2.13.0